feat: add relation error ToStatus func and add user resource role func

+++ relation/del_object.go
@@ -2,11 +2,15 @@ package relation
 
 import (
 	"context"
+	"fmt"
 
 	pb "github.com/ory/keto/proto/ory/keto/relation_tuples/v1alpha2"
 )
 
 func DeleteObjectId(ctx context.Context, namespace, objectId string) error {
+	if writeconn == nil {
+		return ErrWriteConnectNotInitialed
+	}
 	writeClient := pb.NewWriteServiceClient(writeconn)
 	_, err := writeClient.DeleteRelationTuples(ctx, &pb.DeleteRelationTuplesRequest{
 		RelationQuery: &pb.RelationQuery{
@@ -14,5 +18,8 @@ func DeleteObjectId(ctx context.Context, namespace, objectId string) error {
 			Object:    &objectId,
 		},
 	})
-	return err
+	if err != nil {
+		return fmt.Errorf("%w: %w", ErrReadFailed, err)
+	}
+	return nil
 }
